Update import statement for fontkit in README.md
This commit updates the import statement for the `fontkit` library in the README.md file. 

Previously, the import statement was written as `import fontkit from '@pdf-lib/fontkit';`. However, this might cause issues if the library does not have a default export. 

To ensure compatibility with different module systems, the import statement has been changed to `import * as fontkit from '@pdf-lib/fontkit';`. This syntax imports all exports from the `fontkit` library into an object named `fontkit`, which can then be used to access the library's features.